Fix broken license downloads from GitHub
When a license is linked to a file on GitHub, the downloaded HTML will
not be usable. Instead, detect those URLs and download the raw file
instead. This file will usually be a text file and not HTML, so make
sure that the correct extension is used in this case.